Add prettier as a dev dependency (#135)
Previously, the project depended on prettier, but it was not listed
as a dev dependency, so the project was expecting it to be
installed globally, which is not a reasonable expectation. This
commit adds the package as a dev dependency so that the git hooks
defined in husky can run properly without a global dependency.